pyproject.toml
[build-system]
# Defines the tools needed to build your package
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plainbook"
version = "0.0.20"
description = "Plain Language Notebooks"
readme = "README.md"
requires-python = ">=3.8"
license = "BSD-3-Clause"
license-files = ["LICENSE.md"]
authors = [
    {name = "Luca de Alfaro", email = "dealfaro@acm.org"}
]
keywords = ["notebook", "jupyter", "natural language", "data science"]

dependencies = [
    "anthropic",
    "bottle",
    "cherrypy",
    "google-genai",
    "ipykernel",
    "jupyter_client",
    "snapshot-kernel>=0.1.3",
    "nbclient",
    "nbformat",
    "pyyaml",
    "requests",
]

[project.scripts]
plainbook = "plainbook.main:main"

[project.urls]
# Links displayed on the PyPI project page
Homepage = "https://github.com/lucadealfaro/plainbook"
Issues = "https://github.com/lucadealfaro/plainbook/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["plainbook*"]

[tool.setuptools.package-data]
plainbook = ["js/*.js", "css/*.css", "views/*.html", "fonts/*"]
